+++ projects/_03_independent_cqc/_04_model/utils/validate_models.py
@@ -1,3 +1,5 @@
+from datetime import date
+
 import polars as pl
 
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
@@ -15,6 +17,18 @@ non_res_with_dormancy_cols_for_features = [
     IndCQC.related_location,
     IndCQC.time_registered,
     IndCQC.time_since_dormant,
+]
+
+non_res_without_dormancy_cols_for_features = [
+    IndCQC.regulated_activities_offered,
+    IndCQC.cqc_location_import_date,
+    IndCQC.posts_rolling_average_model,
+    IndCQC.services_offered,
+    IndCQC.specialisms_offered,
+    IndCQC.current_rural_urban_indicator_2011,
+    IndCQC.current_region,
+    IndCQC.related_location,
+    IndCQC.time_registered,
 ]
 
 care_home_cols_for_features = [
@@ -52,6 +66,17 @@ def get_expected_row_count_for_model_features(df: pl.DataFrame, model: str) -> i
                 [
                     pl.col(col_for_features).is_not_null()
                     for col_for_features in non_res_with_dormancy_cols_for_features
+                ]
+            ),
+        )
+    elif model == "non_res_without_dormancy_model":
+        df = df.filter(
+            pl.col(IndCQC.care_home) == CareHome.not_care_home,
+            pl.col(IndCQC.cqc_location_import_date) < date(2025, 1, 1),
+            pl.all_horizontal(
+                [
+                    pl.col(col_for_features).is_not_null()
+                    for col_for_features in non_res_without_dormancy_cols_for_features
                 ]
             ),
         )