+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - Polars job [validate_model_01_features](projects/_03_independent_cqc/_04_model/fargate/validate_model_01_features.py) to validate features data. Associated unit tests also added.
   - Updated polars job to validate non_res_with_dormancy_model features and updated associated unit tests.
   - Updated polars job to validate care_home_model features and updated associated unit tests.
+  - Updated polars job to validate non_res_without_dormancy_model features and updated associated unit tests.
 
 ### Changed
 - Remove interim/demo model preprocessing/retraining code.